# frontend/src/shared/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AnalysisProjection": {
        "additionalProperties": false,
        "properties": {
          "analysis_id": {
            "title": "Analysis Id",
            "type": "string"
          },
          "applicability": {
            "$ref": "#/components/schemas/Applicability"
          },
          "canonical_smiles": {
            "title": "Canonical Smiles",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "input_smiles": {
            "title": "Input Smiles",
            "type": "string"
          },
          "policy_snapshot": {
            "additionalProperties": true,
            "title": "Policy Snapshot",
            "type": "object"
          },
          "provenance": {
            "$ref": "#/components/schemas/AnalysisProvenance"
          },
          "requested_endpoints": {
            "items": {
              "enum": [
                "clintox",
                "herg",
                "tox21"
              ],
              "type": "string"
            },
            "title": "Requested Endpoints",
            "type": "array"
          },
          "required_limitations": {
            "items": {
              "type": "string"
            },
            "title": "Required Limitations",
            "type": "array"
          },
          "sections": {
            "$ref": "#/components/schemas/AnalysisSections"
          },
          "served_endpoints": {
            "items": {
              "enum": [
                "clintox",
                "herg",
                "tox21"
              ],
              "type": "string"
            },
            "title": "Served Endpoints",
            "type": "array"
          },
          "unavailable_endpoints": {
            "items": {
              "enum": [
                "clintox",
                "herg",
                "tox21"
              ],
              "type": "string"
            },
            "title": "Unavailable Endpoints",
            "type": "array"
          }
        },
        "required": [
          "input_smiles",
          "canonical_smiles",
          "requested_endpoints",
          "served_endpoints",
          "unavailable_endpoints",
          "sections",
          "applicability",
          "provenance",
          "policy_snapshot",
          "required_limitations",
          "created_at",
          "analysis_id"
        ],
        "title": "AnalysisProjection",
        "type": "object"
      },
      "AnalysisProvenance": {
        "additionalProperties": true,
        "properties": {
          "content_sha256": {
            "title": "Content Sha256",
            "type": "string"
          }
        },
        "required": [
          "content_sha256"
        ],
        "title": "AnalysisProvenance",
        "type": "object"
      },
      "AnalysisSections": {
        "additionalProperties": false,
        "properties": {
          "clintox": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClintoxSection"
              },
              {
                "type": "null"
              }
            ]
          },
          "herg": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HergSection"
              },
              {
                "type": "null"
              }
            ]
          },
          "tox21": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Tox21Section"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "AnalysisSections",
        "type": "object"
      },
      "Applicability": {
        "additionalProperties": true,
        "properties": {
          "method": {
            "title": "Method",
            "type": "string"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "title": "Reasons",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "method",
          "reasons"
        ],
        "title": "Applicability",
        "type": "object"
      },
...
      "AttributionListResponse": {
        "additionalProperties": false,
        "properties": {
          "attributions": {
            "items": {
              "$ref": "#/components/schemas/AttributionProjection"
            },
            "title": "Attributions",
            "type": "array"
          }
        },
        "required": [
          "attributions"
        ],
        "title": "AttributionListResponse",
        "type": "object"
      },
      "AttributionProjection": {
        "additionalProperties": false,
        "properties": {
          "analysis_id": {
            "title": "Analysis Id",
            "type": "string"
          },
          "content_sha256": {
            "title": "Content Sha256",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "endpoint": {
            "enum": [
              "clintox",
              "herg",
              "tox21"
            ],
            "title": "Endpoint",
            "type": "string"
          },
          "method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Method"
          },
          "model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Id"
          },
          "observation_id": {
            "title": "Observation Id",
            "type": "string"
          },
          "required_limitations": {
            "items": {
              "type": "string"
            },
            "title": "Required Limitations",
            "type": "array"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "completed",
              "partial"
            ],
            "title": "Status",
            "type": "string"
          },
          "task": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task"
          },
          "top_tokens": {
            "items": {
              "$ref": "#/components/schemas/AttributionToken"
            },
            "title": "Top Tokens",
            "type": "array"
          }
        },
        "required": [
          "observation_id",
          "run_id",
          "created_at",
          "content_sha256",
          "required_limitations",
          "analysis_id",
          "endpoint",
          "task",
          "status",
          "method",
          "model_id",
          "top_tokens"
        ],
        "title": "AttributionProjection",
        "type": "object"
      },
...
      "EvidenceListResponse": {
        "additionalProperties": false,
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceRecordView"
            },
            "title": "Evidence",
            "type": "array"
          }
        },
        "required": [
          "evidence",
          "count"
        ],
        "title": "EvidenceListResponse",
        "type": "object"
      },
      "EvidenceRecordView": {
        "additionalProperties": false,
        "description": "Bounded, normalised external material. ``abstract_or_excerpt`` is\nuntrusted content, never instructions.",
        "properties": {
          "abstract_or_excerpt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Abstract Or Excerpt"
          },
          "authors": {
            "items": {
              "type": "string"
            },
            "title": "Authors",
            "type": "array"
          },
          "canonical_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Url"
          },
          "content_sha256": {
            "title": "Content Sha256",
            "type": "string"
          },
          "evidence_id": {
            "title": "Evidence Id",
            "type": "string"
          },
          "identifier": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Identifier",
            "type": "object"
          },
          "normalized_facts": {
            "additionalProperties": true,
            "title": "Normalized Facts",
            "type": "object"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "rejection_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejection Reason"
          },
          "retrieved_at": {
            "title": "Retrieved At",
            "type": "string"
          },
          "source_quality_tier": {
            "title": "Source Quality Tier",
            "type": "string"
          },
          "source_type": {
            "title": "Source Type",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "untrusted_external_content": {
            "const": true,
            "title": "Untrusted External Content",
            "type": "boolean"
          }
        },
        "required": [
          "evidence_id",
          "title",
          "authors",
          "published_at",
          "source_type",
          "source_quality_tier",
          "identifier",
          "canonical_url",
          "abstract_or_excerpt",
          "normalized_facts",
          "status",
          "rejection_reason",
          "provider",
          "retrieved_at",
          "content_sha256",
          "untrusted_external_content"
        ],
        "title": "EvidenceRecordView",
        "type": "object"
      },
...
      "GroundedAnswer": {
        "additionalProperties": false,
        "properties": {
          "answer_id": {
            "title": "Answer Id",
            "type": "string"
          },
          "answer_markdown": {
            "title": "Answer Markdown",
            "type": "string"
          },
          "candidate_generation": {
            "title": "Candidate Generation",
            "type": "integer"
          },
          "claims": {
            "items": {
              "$ref": "#/components/schemas/Claim"
            },
            "title": "Claims",
            "type": "array"
          },
          "content_sha256": {
            "title": "Content Sha256",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "is_fallback": {
            "title": "Is Fallback",
            "type": "boolean"
          },
          "limitations": {
            "items": {
              "$ref": "#/components/schemas/Limitation"
            },
            "title": "Limitations",
            "type": "array"
          },
          "recommended_next_steps": {
            "items": {
              "$ref": "#/components/schemas/RecommendedNextStep"
            },
            "title": "Recommended Next Steps",
            "type": "array"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "schema_version",
          "answer_id",
          "run_id",
          "answer_markdown",
          "claims",
          "limitations",
          "recommended_next_steps",
          "candidate_generation",
          "is_fallback",
          "content_sha256",
          "created_at"
        ],
        "title": "GroundedAnswer",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Limitation": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "enum": [
              "uncalibrated_probability",
              "applicability_is_rule_based",
              "attribution_not_causality",
              "endpoint_unavailable",
              "evidence_scope_limited",
              "screening_not_safety_assessment"
            ],
            "title": "Code",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "code",
          "text"
        ],
        "title": "Limitation",
        "type": "object"
      },
...
      "ObservationResponse": {
        "additionalProperties": false,
        "properties": {
          "canonical_payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Payload"
          },
          "content_sha256": {
            "title": "Content Sha256",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "model_projection": {
            "additionalProperties": true,
            "title": "Model Projection",
            "type": "object"
          },
          "observation_id": {
            "title": "Observation Id",
            "type": "string"
          },
          "producer": {
            "title": "Producer",
            "type": "string"
          },
          "provenance": {
            "additionalProperties": true,
            "title": "Provenance",
            "type": "object"
          },
          "required_limitations": {
            "items": {
              "type": "string"
            },
            "title": "Required Limitations",
            "type": "array"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "observation_id",
          "run_id",
          "producer",
          "kind",
          "schema_version",
          "model_projection",
          "provenance",
          "required_limitations",
          "content_sha256",
          "created_at"
        ],
        "title": "ObservationResponse",
        "type": "object"
      },
...
      "RecommendedNextStep": {
        "additionalProperties": false,
        "properties": {
          "basis_claim_ids": {
            "items": {
              "type": "string"
            },
            "title": "Basis Claim Ids",
            "type": "array"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text",
          "basis_claim_ids"
        ],
        "title": "RecommendedNextStep",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "ToxAgent control plane.\n\nEvidence and decision support over the ToxPred predictor. Every number in an\naccepted answer resolves to a stored observation and a field path; hERG, Tox21\nand ClinTox stay three separate measurements; and no schema here can carry an\naggregate toxicity or safety verdict.\n",
    "title": "ToxAgent control plane",
    "version": "0.1.0.dev0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/sessions/{session_id}/analyses/{analysis_id}": {
      "get": {
        "operationId": "get_analysis_v1_sessions__session_id__analyses__analysis_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "title": "Analysis Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include_raw",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Raw",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisProjection"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Analysis",
        "tags": [
          "toxagent"
        ]
      }
    },
    "/v1/sessions/{session_id}/analyses/{analysis_id}/attributions": {
      "get": {
        "description": "List bounded attribution observations for one immutable analysis.\n\nThe projection is intentionally the same top-token view supplied to the\nmodel, rather than canonical/raw provider output. An attribution belongs\nto exactly one endpoint (and one Tox21 task when applicable), so the UI\ncannot construct an aggregate explanation from this endpoint.",
        "operationId": "list_attributions_v1_sessions__session_id__analyses__analysis_id__attributions_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "title": "Analysis Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttributionListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Attributions",
        "tags": [
          "toxagent"
        ]
      }
    },
    "/v1/sessions/{session_id}/answers/{answer_id}": {
      "get": {
        "operationId": "get_answer_v1_sessions__session_id__answers__answer_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "answer_id",
            "required": true,
            "schema": {
              "title": "Answer Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroundedAnswer"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Answer",
        "tags": [
          "toxagent"
        ]
      }
    },
...
    "/v1/sessions/{session_id}/evidence": {
      "get": {
        "operationId": "list_evidence_v1_sessions__session_id__evidence_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "accepted",
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Evidence",
        "tags": [
          "toxagent"
        ]
      }
    },
    "/v1/sessions/{session_id}/evidence/{evidence_id}": {
      "get": {
        "description": "Return the bounded, normalized evidence projection for its owner.\n\nThis deliberately mirrors the model-visible view, plus audit-safe\ntransport metadata. ``raw_payload_ref`` remains object-store/auditor\nmaterial (W4-09) and is never a browser URL or a model capability.",
        "operationId": "get_evidence_v1_sessions__session_id__evidence__evidence_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "evidence_id",
            "required": true,
            "schema": {
              "title": "Evidence Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceRecordView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Evidence",
        "tags": [
          "toxagent"
        ]
      }
    },
...
    "/v1/sessions/{session_id}/observations/{observation_id}": {
      "get": {
        "description": "The other end of every claim's ``observation_id`` (plan section 5.5).\n\nWithout this, ``field_path`` and ``source_value`` on a claim are citations\nto nothing a client can open. The lossless ``canonical_payload`` stays\naudit-only, same gate as ``analyses`` ``include_raw``: a claim only ever\nneeded the bounded ``model_projection`` to be valid.",
        "operationId": "get_observation_v1_sessions__session_id__observations__observation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "observation_id",
            "required": true,
            "schema": {
              "title": "Observation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObservationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Observation",
        "tags": [
          "toxagent"
        ]
      }
    },
...
  }
}